# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"Data Aggregator API","description":"Data extraction and aggregation tool with FSM stage orchestration","version":"0.1.0"},"servers":[{"url":"/api/dat"}],"paths":{"/v1/runs":{"post":{"tags":["dat"],"summary":"Create Run","description":"Create a new DAT run.","operationId":"create_run_v1_runs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRunRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRunResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["dat"],"summary":"List Runs","description":"List DAT runs.","operationId":"list_runs_v1_runs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}":{"get":{"tags":["dat"],"summary":"Get Run","description":"Get DAT run details.","operationId":"get_run_v1_runs__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/{stage}":{"get":{"tags":["dat"],"summary":"Get Stage Status","description":"Get stage status.","operationId":"get_stage_status_v1_runs__run_id__stages__stage__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}},{"name":"stage","in":"path","required":true,"schema":{"type":"string","title":"Stage"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/discovery/lock":{"post":{"tags":["dat"],"summary":"Lock Discovery","description":"Lock discovery stage - scan for files in a folder.\n\nPer ADR-0004: Discovery is the first stage and must be locked before Selection.","operationId":"lock_discovery_v1_runs__run_id__stages_discovery_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/selection/scan":{"post":{"tags":["dat"],"summary":"Scan Selection","description":"Scan folder for files without locking stage.","operationId":"scan_selection_v1_runs__run_id__stages_selection_scan_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/selection/lock":{"post":{"tags":["dat"],"summary":"Lock Selection","description":"Lock selection stage - finalize file selection.","operationId":"lock_selection_v1_runs__run_id__stages_selection_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SelectionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/context/lock":{"post":{"tags":["dat"],"summary":"Lock Context","description":"Lock context stage - set profile and aggregation levels.","operationId":"lock_context_v1_runs__run_id__stages_context_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"type":"object","additionalProperties":true},{"type":"null"}],"title":"Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/table_selection/lock":{"post":{"tags":["dat"],"summary":"Lock Table Selection","description":"Lock table selection stage.","operationId":"lock_table_selection_v1_runs__run_id__stages_table_selection_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableSelectionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/parse/lock":{"post":{"tags":["dat"],"summary":"Lock Parse","description":"Lock parse stage - start data extraction.","operationId":"lock_parse_v1_runs__run_id__stages_parse_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParseRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/parse/cancel":{"post":{"tags":["dat"],"summary":"Cancel Parse","description":"Cancel ongoing parse operation.\n\nPer ADR-0014: Cancellation events are logged for audit.","operationId":"cancel_parse_v1_runs__run_id__stages_parse_cancel_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}},{"name":"reason","in":"query","required":false,"schema":{"type":"string","default":"user_requested","title":"Reason"}},{"name":"actor","in":"query","required":false,"schema":{"type":"string","default":"user","title":"Actor"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/export":{"post":{"tags":["dat"],"summary":"Export Dataset","description":"Export run as DataSet.","operationId":"export_dataset_v1_runs__run_id__export_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/profiles":{"get":{"tags":["dat"],"summary":"List Profiles","description":"List available profiles.","operationId":"list_profiles_v1_profiles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/runs/{run_id}/stages/table_availability/scan":{"get":{"tags":["dat"],"summary":"Scan Table Availability","description":"Scan for available tables from selected files.\n\nPer ADR-0008: Table availability must show actual row/column counts.","operationId":"scan_table_availability_v1_runs__run_id__stages_table_availability_scan_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/table_availability/lock":{"post":{"tags":["dat"],"summary":"Lock Table Availability","description":"Lock table availability stage with discovered tables.\n\nPer ADR-0008: Table availability must show actual row/column counts.","operationId":"lock_table_availability_v1_runs__run_id__stages_table_availability_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/table_selection/tables":{"get":{"tags":["dat"],"summary":"Get Table Selection Tables","description":"Get available tables for selection.","operationId":"get_table_selection_tables_v1_runs__run_id__stages_table_selection_tables_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/{stage}/unlock":{"post":{"tags":["dat"],"summary":"Unlock Stage","description":"Unlock a stage with artifact preservation per ADR-0002.","operationId":"unlock_stage_v1_runs__run_id__stages__stage__unlock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}},{"name":"stage","in":"path","required":true,"schema":{"type":"string","title":"Stage"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/preview/lock":{"post":{"tags":["dat"],"summary":"Lock Preview","description":"Lock preview stage with generated preview data from selected tables.\n\nPer ADR-0004: Preview is an optional stage BEFORE Parse.\nIt shows a preview of selected tables to verify data before parsing.","operationId":"lock_preview_v1_runs__run_id__stages_preview_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/preview/data":{"get":{"tags":["dat"],"summary":"Get Preview","description":"Get preview of parsed data from preview stage artifact.","operationId":"get_preview_v1_runs__run_id__stages_preview_data_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}},{"name":"rows","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Rows"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/export/lock":{"post":{"tags":["dat"],"summary":"Lock Export","description":"Lock export stage with dataset generation.","operationId":"lock_export_v1_runs__run_id__stages_export_lock_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/export/summary":{"get":{"tags":["dat"],"summary":"Get Export Summary","description":"Get export summary before executing export.","operationId":"get_export_summary_v1_runs__run_id__stages_export_summary_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/export/execute":{"post":{"tags":["dat"],"summary":"Execute Export Stage","description":"Execute export stage.","operationId":"execute_export_stage_v1_runs__run_id__stages_export_execute_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/parse/progress":{"get":{"tags":["dat"],"summary":"Get Parse Progress","description":"Get parse stage progress.","operationId":"get_parse_progress_v1_runs__run_id__stages_parse_progress_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/runs/{run_id}/stages/parse/start":{"post":{"tags":["dat"],"summary":"Start Parse","description":"Start parse stage execution with table selection data.","operationId":"start_parse_v1_runs__run_id__stages_parse_start_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs":{"post":{"tags":["jobs","jobs"],"summary":"Submit a new job","description":"Submit a new background job for processing.","operationId":"submit_job_jobs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobSubmission"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BackgroundJob"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["jobs","jobs"],"summary":"List jobs","description":"List jobs with optional filtering and pagination.","operationId":"list_jobs_jobs_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/JobStatus"},{"type":"null"}],"description":"Filter by status","title":"Status"},"description":"Filter by status"},{"name":"run_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by run ID","title":"Run Id"},"description":"Filter by run ID"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Max jobs to return","default":50,"title":"Limit"},"description":"Max jobs to return"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Pagination offset","default":0,"title":"Offset"},"description":"Pagination offset"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/queue/status":{"get":{"tags":["jobs","jobs"],"summary":"Get queue status","description":"Get current status of the job queue.","operationId":"get_queue_status_jobs_queue_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobQueueStatus"}}}}}}},"/jobs/{job_id}":{"get":{"tags":["jobs","jobs"],"summary":"Get job status","description":"Get the current status of a specific job.","operationId":"get_job_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BackgroundJob"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["jobs","jobs"],"summary":"Cancel or delete job","description":"Cancel a running/pending job or delete a completed job.","operationId":"cancel_or_delete_job_jobs__job_id__delete","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}},{"name":"delete","in":"query","required":false,"schema":{"type":"boolean","description":"Delete instead of cancel","default":false,"title":"Delete"},"description":"Delete instead of cancel"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BackgroundJob"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BackgroundJob":{"properties":{"job_id":{"type":"string","pattern":"^job_[a-f0-9]{16}$","title":"Job Id","description":"Unique job identifier"},"job_type":{"$ref":"#/components/schemas/JobType","description":"Type of job"},"status":{"$ref":"#/components/schemas/JobStatus","description":"Current job status","default":"pending"},"progress":{"$ref":"#/components/schemas/JobProgress","description":"Progress information"},"result":{"anyOf":[{"$ref":"#/components/schemas/JobResult"},{"type":"null"}],"description":"Job result when completed"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"When job was created"},"started_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Started At","description":"When job started processing"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At","description":"When job finished"},"run_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id","description":"Associated DAT run ID"},"file_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Path","description":"Relative path to input file"},"options":{"additionalProperties":true,"type":"object","title":"Options","description":"Job-specific options"},"priority":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Priority","description":"Job priority (higher = more urgent)","default":0},"retries":{"type":"integer","minimum":0.0,"title":"Retries","description":"Number of retry attempts","default":0},"max_retries":{"type":"integer","maximum":10.0,"minimum":0.0,"title":"Max Retries","description":"Maximum retry attempts","default":3},"timeout_seconds":{"type":"integer","maximum":86400.0,"minimum":60.0,"title":"Timeout Seconds","description":"Job timeout in seconds","default":3600},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error","description":"Last error message if failed"}},"type":"object","required":["job_id","job_type","created_at"],"title":"BackgroundJob","description":"Background job model.\n\nRepresents a long-running job for processing large files.\n\nAttributes:\n    job_id: Unique job identifier.\n    job_type: Type of job.\n    status: Current job status.\n    progress: Progress information.\n    result: Job result (when completed).\n    created_at: When job was created.\n    started_at: When job started processing.\n    completed_at: When job finished.\n    run_id: Associated DAT run ID.\n    file_path: Path to input file.\n    options: Job-specific options.\n    priority: Job priority (higher = more urgent).\n    retries: Number of retry attempts.\n    max_retries: Maximum retry attempts.\n    timeout_seconds: Job timeout.\n    error: Last error if failed."},"CreateRunRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"profile_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Id"}},"type":"object","title":"CreateRunRequest","description":"Request to create a new DAT run."},"CreateRunResponse":{"properties":{"run_id":{"type":"string","title":"Run Id"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"profile_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Id"}},"type":"object","required":["run_id","name","created_at"],"title":"CreateRunResponse","description":"Response for created run."},"ExportRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"aggregation_levels":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Aggregation Levels"}},"type":"object","title":"ExportRequest","description":"Request to export as DataSet."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobListResponse":{"properties":{"jobs":{"items":{"$ref":"#/components/schemas/BackgroundJob"},"type":"array","title":"Jobs","description":"List of jobs"},"total":{"type":"integer","minimum":0.0,"title":"Total","description":"Total matching jobs"},"limit":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Limit","description":"Max jobs returned","default":50},"offset":{"type":"integer","minimum":0.0,"title":"Offset","description":"Pagination offset","default":0}},"type":"object","required":["jobs","total"],"title":"JobListResponse","description":"Response for listing jobs.\n\nAttributes:\n    jobs: List of jobs.\n    total: Total number of jobs matching query.\n    limit: Maximum jobs returned.\n    offset: Offset for pagination."},"JobProgress":{"properties":{"percentage":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Percentage","description":"Completion percentage","default":0.0},"current_step":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Current Step","description":"Current step description"},"total_steps":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Total Steps","description":"Total number of steps"},"current_step_number":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Current Step Number","description":"Current step number"},"rows_processed":{"type":"integer","minimum":0.0,"title":"Rows Processed","description":"Number of rows processed","default":0},"total_rows":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total Rows","description":"Total rows to process"},"bytes_processed":{"type":"integer","minimum":0.0,"title":"Bytes Processed","description":"Bytes processed","default":0},"total_bytes":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total Bytes","description":"Total bytes to process"},"eta_seconds":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Eta Seconds","description":"Estimated time to completion in seconds"}},"type":"object","title":"JobProgress","description":"Progress information for a job.\n\nAttributes:\n    percentage: Completion percentage (0-100).\n    current_step: Current step description.\n    total_steps: Total number of steps.\n    current_step_number: Current step number.\n    rows_processed: Number of rows processed.\n    total_rows: Total rows to process.\n    bytes_processed: Bytes processed.\n    total_bytes: Total bytes to process.\n    eta_seconds: Estimated time to completion."},"JobQueueStatus":{"properties":{"pending_count":{"type":"integer","minimum":0.0,"title":"Pending Count","description":"Pending jobs","default":0},"running_count":{"type":"integer","minimum":0.0,"title":"Running Count","description":"Running jobs","default":0},"completed_today":{"type":"integer","minimum":0.0,"title":"Completed Today","description":"Completed today","default":0},"failed_today":{"type":"integer","minimum":0.0,"title":"Failed Today","description":"Failed today","default":0},"max_concurrent":{"type":"integer","minimum":1.0,"title":"Max Concurrent","description":"Max concurrent jobs","default":4},"queue_healthy":{"type":"boolean","title":"Queue Healthy","description":"Queue health status","default":true}},"type":"object","title":"JobQueueStatus","description":"Status of the job queue.\n\nAttributes:\n    pending_count: Number of pending jobs.\n    running_count: Number of running jobs.\n    completed_today: Jobs completed today.\n    failed_today: Jobs failed today.\n    max_concurrent: Maximum concurrent jobs allowed.\n    queue_healthy: Whether queue is healthy."},"JobResult":{"properties":{"success":{"type":"boolean","title":"Success","description":"Whether job completed successfully"},"output_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output Path","description":"Relative path to output artifact"},"row_count":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Row Count","description":"Number of rows in output"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message","description":"Error message if failed"},"warnings":{"items":{"type":"string"},"type":"array","title":"Warnings","description":"Warning messages"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata","description":"Additional result metadata"}},"type":"object","required":["success"],"title":"JobResult","description":"Result of a completed job.\n\nAttributes:\n    success: Whether job completed successfully.\n    output_path: Path to output file/artifact.\n    row_count: Number of rows in output.\n    error_message: Error message if failed.\n    warnings: Warning messages during processing.\n    metadata: Additional result metadata."},"JobStatus":{"type":"string","enum":["pending","running","completed","failed","cancelled"],"title":"JobStatus","description":"Status of a background job."},"JobSubmission":{"properties":{"job_type":{"$ref":"#/components/schemas/JobType","description":"Type of job to run"},"run_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id","description":"Associated DAT run ID"},"file_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Path","description":"Relative path to input file"},"options":{"additionalProperties":true,"type":"object","title":"Options","description":"Job-specific options"},"priority":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Priority","description":"Job priority","default":0},"timeout_seconds":{"anyOf":[{"type":"integer","maximum":86400.0,"minimum":60.0},{"type":"null"}],"title":"Timeout Seconds","description":"Custom timeout"}},"type":"object","required":["job_type"],"title":"JobSubmission","description":"Request to submit a new job.\n\nAttributes:\n    job_type: Type of job to run.\n    run_id: Associated DAT run ID.\n    file_path: Path to input file.\n    options: Job-specific options.\n    priority: Job priority.\n    timeout_seconds: Custom timeout."},"JobType":{"type":"string","enum":["parse","export","validate","transform"],"title":"JobType","description":"Type of background job."},"ParseRequest":{"properties":{"column_mappings":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Column Mappings"}},"type":"object","title":"ParseRequest","description":"Request to start parse stage."},"ScanRequest":{"properties":{"folder_path":{"type":"string","title":"Folder Path"},"recursive":{"type":"boolean","title":"Recursive","default":true}},"type":"object","required":["folder_path"],"title":"ScanRequest","description":"Request to scan a folder for files."},"SelectionRequest":{"properties":{"source_paths":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Source Paths"},"selected_files":{"items":{"type":"string"},"type":"array","title":"Selected Files"},"recursive":{"type":"boolean","title":"Recursive","default":true}},"type":"object","required":["selected_files"],"title":"SelectionRequest","description":"Request to lock selection stage."},"TableSelectionRequest":{"properties":{"selected_tables":{"additionalProperties":{"items":{"type":"string"},"type":"array"},"type":"object","title":"Selected Tables"}},"type":"object","required":["selected_tables"],"title":"TableSelectionRequest","description":"Request to lock table selection stage."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}